New: Persist the result items
Writing the result items after a successfull crawl to the database.
We create only one entry for the combination of the url and origin url.
We also keep this items in sync. So we don`t create new items when we
check the urls again. We just update the status and the checked at date.