Breaking: Change Vendor name to Unikka
The package was published as noedisch package before but is not available anymore.
Now we adjust the vendor name and name space to fit the unikka eco system.

The package was authored by me under the MIT license and so we continue now
as an unikka package.

BREAKING CHANGE: Vendor name and namespace change